project/cuervos/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.9.2
python-multipart==0.0.9
python-socketio==5.11.0